Change to single-line error and warning messages.
It was a nice idea to use multi-line error messages to make them
really obvious, but unfortunately it is terrible in practice.  Often
errors will occur on multiple nodes simultaneously and the error
messages will be interleaved and virtually unreadable.

I also prefixed each message with "ior", to make it clear that
it is ior that produced the message, and not the scheduler, MPI, or
something else.