Setup googletest for unit testing
This adds the googletest unit test framework. It is linked statically,
compiled and run only when 'make check' is called. Googletest library
lives in a new folder 'libs' and is added as a git submodule form the
github repo https://github.com/google/googletest
Test are created and compiled in a new folder 'tests' to keep them
separate for release path.
Initially a test for 'getOffsetArraySequential' is implemented.
This is just a starting point to get more confidence in the code when
changing or adding features as i have these when touching core parts of
IOR.