Travis-CI: Add config file
This adds a .travis.yaml file to use travis-ci for continuous
integration.
It currently compiles IOR on an ubuntu 14.04 machine with
only the POSIX, MPIIO and HDF5 backends enabled. For Lustre, GPFS,
NCMPI, HDFS and S3 libraries needs to be compiled as they are not
present in the ubuntu repositories. Lustre and NCMPI should be easy when
ubuntu 16.04 is available at travis-ci. S3 could be compiled by source
which shouldn't be hard but needs to be done if one is interested.
Hadoop needs a more complex installation process. GPFS seams not
possible as it is proprietary and needs a license as far as i know.

Repo Owner needs to enable CI on travis-ci.org. It's free for open
source git hub repos.