export generic posix-based backend calls
aiori internally has members for statfs, mkdir, rmdir, access,
and stat calls that are set to generic variants that assume a
POSIX-based FS if they aren't provided by the backend
implementation. this commit explicitly exports those calls,
renames them to indicate they assume POSIX, and forces all current
POSIX-based backends to manually opt in to them.

aiori_select was additionally updated to assume a POSIX-based
backend if the implementation for these calls aren't provided,
with a warning printed out to alert the user of this assumption.