Fix misdocumented IOR END for IOR STOP.
The prose mentions IOR END, which is not actually a thing, as I gather from src/parse_options.c.
Also, the words are case-insensitive, which might be welcomed by some people preferring
lowercase aesthetics;-)  If they are treated that way (strncasecmp()), they should be documented
that way. For brevity, I did not repeat that for RUN, but one could do that, too.