userDoc: fix mentioning of IOR STOP and be more specific about case (#457)
* Fix misdocumented IOR END for IOR STOP.
The prose mentions IOR END, which is not actually a thing, as I gather from src/parse_options.c.
Also, the words are case-insensitive, which might be welcomed by some people preferring
lowercase aesthetics;-)  If they are treated that way (strncasecmp()), they should be documented
that way. For brevity, I did not repeat that for RUN, but one could do that, too.
* Explicitly clarify parser behaviour also for RUN.
All of the control words are compared via strncasecmp(). Be vocal about that. The documentation
should not be pretty, but precise.
* fix syntax and grammar

A space too much messes up the rst list, and a dot each was missing for English grammar.