ior: improve error for existing output file
If the '-E' option is not used and existing output files are found,
the current error message says that the files will be "overwritten"
but this is not totally accurate.  Without '-E' the file will be
deleted first, which makes a difference if the files were precreated
with a specific layout.

Update the error to report that files will be deleted in this case.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>